refactor(editor): delete dogfooded flag paths (isVirtualized, isMobileBar)
The recursive OutlineNode render fallback and the mobile-bar rollback flag
were delete-after-dogfooding (ADR 0019 / ADR 0030); dogfooding is done.
NodeCommands moves to node-commands.ts; the now-dead useSelectionEdge,
.outline-children CSS, and drag-collapsed toggles go with the path. The
mirrors flag-OFF e2e parity tests covered the deleted baseline and are
removed with it (isMirrorsEnabled itself stays).

Closes #160
