+++ src/data/selection-state.ts
@@ -24,12 +24,11 @@ import { getViewIsHidden, getViewRootId } from "./view-state";
  * Shape: the actor is a MODULE SINGLETON (`createActor(...).start()`), NOT a
  * `useMachine` -- it's mirrored like {@link view-state} and the tree store. The
  * stable command/keyboard closures read it live at event time (`getSnapshot`),
- * and each row subscribes to its OWN slice via {@link useSelectionEdge} (now
- * `@xstate/react`'s `useSelector` with an edge-equality compare), so a selection
- * change re-renders only the rows entering or leaving it -- preserving ADR 0014's
- * per-node-render budget. Selection is NEVER threaded as a prop. The public API
- * of this module is unchanged from the pre-XState version; only the internals
- * moved into the machine, so a rollback is "restore the old file + drop two deps".
+ * and each row subscribes to its OWN slice via {@link "./selection-fill"}'s
+ * `useSelectionFill` (a `useSyncExternalStore` keyed by row, fed by
+ * {@link subscribeSelection}), so a selection change re-renders only the rows
+ * entering or leaving it -- preserving ADR 0014's per-node-render budget.
+ * Selection is NEVER threaded as a prop.
  *
  * Model: a selection is a **contiguous run of siblings under one parent**
  * (`rootIds`), and selecting a node implies its whole subtree (you select roots;
@@ -159,37 +158,6 @@ function computeAllInView(): SelectionData | null {
   const sibs = visibleSiblings(index, getViewRootId());
   if (sibs.length === 0) return null;
   return computeRange(sibs[0]!.id, sibs[sibs.length - 1]!.id, index);
-}
-
-/** id -> edge map for one selection, built once and reused for every row. Keyed
- *  on the `SelectionData` reference (stable per selection -- a no-op transition
- *  preserves it), so it rebuilds only when the selection actually changes, and a
- *  replaced `SelectionData` is GC'd with its map. This keeps `edgeOf` O(1) per
- *  row instead of an O(rootIds) `indexOf` on every subscribed row per snapshot
- *  (the pre-XState `recomputeEdges` behavior). */
-const edgeMaps = new WeakMap<SelectionData, Map<string, SelectionEdge>>();
-function edgeMapFor(data: SelectionData): Map<string, SelectionEdge> {
-  let m = edgeMaps.get(data);
-  if (m) return m;
-  m = new Map();
-  const ids = data.rootIds;
-  if (ids.length === 1) {
-    m.set(ids[0]!, "single");
-  } else {
-    m.set(ids[0]!, "top");
-    m.set(ids[ids.length - 1]!, "bottom");
-    for (let i = 1; i < ids.length - 1; i++) m.set(ids[i]!, "middle");
-  }
-  edgeMaps.set(data, m);
-  return m;
-}
-
-/** The slab edge for a node id given the current selection, or null when it
- *  isn't a selected root. O(1): consults the memoized {@link edgeMapFor} map, so
- *  the `useSelector` selector stays cheap across every subscribed row. */
-function edgeOf(id: string, data: SelectionData | null): SelectionEdge | null {
-  if (!data) return null;
-  return edgeMapFor(data).get(id) ?? null;
 }
 
 // --- the machine ------------------------------------------------------------
@@ -349,22 +317,6 @@ export function clearSelection() {
   selectionActor.send({ type: "clear" });
 }
 
-/**
- * Per-node subscription to this node's slab edge: a row re-renders only when ITS
- * edge changes (entering/leaving the selection or shifting top<->middle<->bottom),
- * never on an unrelated selection change. `useSelector` runs the selector on every
- * snapshot but only re-renders when the returned edge value differs (`Object.is`),
- * which is what keeps multi-selection inside ADR 0014's per-node-render budget.
- * Returns null when the node isn't a selected root.
- */
-export function useSelectionEdge(id: string): SelectionEdge | null {
-  return useSelector(
-    selectionActor,
-    (snap) => edgeOf(id, snap.context.data),
-    Object.is,
-  );
-}
-
 /** Reactive `rootIds` for the actions menu. `useSelector` re-renders only when the
  *  array value changes (`Object.is`); the context `rootIds` reference is stable
  *  across no-op transitions, so this never churns on an unrelated event. */
@@ -387,9 +339,9 @@ export function useIsSelectionActive(): boolean {
  *  {@link "./selection-fill"}'s module-singleton mirror (ADR 0019's windowed
  *  list has no DOM nesting for a root's tint to paint behind its descendants,
  *  so the fill map needs its own walk -- this is how it learns a selection
- *  changed without itself being a React subscriber). Not part of the frozen
- *  consumer-facing API (selectSingle/extendSelection/useSelectionEdge/...) --
- *  an internal seam for the one other module in this slice. */
+ *  changed without itself being a React subscriber). Not part of the
+ *  consumer-facing API (selectSingle/extendSelection/...) -- an internal seam
+ *  for the one other module in this slice. */
 export function subscribeSelection(cb: () => void): () => void {
   const sub = selectionActor.subscribe(cb);
   return () => sub.unsubscribe();
