+++ src/plugins/route-bible/passage-edit-popover.tsx
@@ -17,7 +17,7 @@ import {
 } from "react";
 import { createPortal } from "react-dom";
 
-import type { NodeCommands } from "../../components/OutlineNode";
+import type { NodeCommands } from "../../components/node-commands";
 import type { PluginContext } from "../types";
 
 import { placeCaretAtEnd } from "../../components/caret-place";
