+++ src/plugins/types.ts
@@ -12,7 +12,7 @@ import type { ComponentType, ReactNode } from "react";
 
 // Type-only (erased at runtime) -- PluginContext.mutations IS the promoted
 // NodeCommands (D8), so we reference its type without a runtime import cycle.
-import type { NodeCommands } from "../components/OutlineNode";
+import type { NodeCommands } from "../components/node-commands";
 import type { FilterOperator } from "../data/filter-query";
 import type { Node, TreeIndex } from "../data/tree";
